# specs/1-customer-success-fte/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Customer Success Digital FTE API
  description: API for Customer Success Digital FTE - Multi-channel customer support AI agent
  version: 1.0.0
  contact:
    name: Support Team
    email: support@example.com
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
  - url: https://api.example.com/api/v1
    description: Production server
...
paths:
  /tickets:
    post:
      summary: Create a new support ticket
      operationId: createTicket
      tags:
        - Tickets
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTicketRequest'
      responses:
        '201':
          description: Ticket created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketResponse'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    get:
      summary: List all tickets
      operationId: listTickets
      tags:
        - Tickets
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [open, in_progress, pending_customer, escalated, resolved, closed]
        - name: channel
          in: query
          schema:
            type: string
            enum: [email, whatsapp, web_form]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        '200':
          description: List of tickets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TicketResponse'
  /tickets/{ticket_id}:
    get:
      summary: Get ticket by ID
      operationId: getTicketById
      tags:
        - Tickets
      parameters:
        - name: ticket_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Ticket details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketResponse'
        '404':
          description: Ticket not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    CreateTicketRequest:
      type: object
      required:
        - customer_email
        - channel
        - description
      properties:
        customer_email:
          type: string
          format: email
        channel:
          type: string
          enum: [email, whatsapp, web_form]
        subject:
          type: string
          maxLength: 500
        description:
          type: string
          minLength: 1
        priority:
          type: string
          enum: [low, normal, high, urgent]
          default: normal
        metadata:
          type: object
    TicketResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        ticket_number:
          type: string
          example: TKT-2026-000123
        customer_id:
          type: string
          format: uuid
        channel:
          type: string
          enum: [email, whatsapp, web_form]
        subject:
          type: string
        description:
          type: string
        priority:
          type: string
          enum: [low, normal, high, urgent]
        status:
          type: string
          enum: [open, in_progress, pending_customer, escalated, resolved, closed]
        sentiment_score:
          type: number
          format: float
          minimum: -1.0
          maximum: 1.0
        created_at:
          type: string
          format: date-time
        first_response_at:
          type: string
          format: date-time
        resolved_at:
          type: string
          format: date-time
...
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - ApiKeyAuth: []
  - BearerAuth: []